requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
certifi==2020.6.20        # via requests
cfg-load==0.9.0           # via edapy (setup.py)
chardet==3.0.4            # via requests
click==7.1.2              # via edapy (setup.py)
idna==2.10                # via requests
mpu[io]==0.23.1           # via cfg-load
numpy==1.19.1             # via pandas
pandas==1.0.5             # via edapy (setup.py)
pillow==7.2.0             # via edapy (setup.py)
pypdf2==1.26.0            # via edapy (setup.py)
python-dateutil==2.8.1    # via pandas
pytz==2020.1              # via cfg-load, mpu, pandas
pyyaml==5.3.1             # via cfg-load, edapy (setup.py)
requests==2.24.0          # via cfg-load
six==1.15.0               # via cfg-load, python-dateutil
tzlocal==2.1              # via mpu
urllib3==1.25.10          # via requests
